Button select error fix.
The compiler warning fix led to an issue where, when clicking the same
button (thus unselecting it, changing its value to '') would cause the
code to break, because due to the asynchronous nature of the useEffects,
the empty string would be passed to the solver.

So I just done away with one of the useEffects and also the
'selectWarning' state variable.