Credits message below main container.
Had to change the results <div> a bit because, for some reason, adding
flex-direction to the 'home' CSS class made the results center things.

Now it only centers (vertically) when there's no instructions. It looks
better this way, whereas it looks better the old way when there are
instructions displayed.