Actually solving the Symbol.name bug.
The reason for 'Symbol.name' being 'undefined' on ocasion was because,
whenever there was either a cube, sphere or tetrahedron in 'finalSymbols',
they would return as undefined due to not being present in 'compoundIcon'
(because they shouldn't ever return as any of those 3, to begin with).

Now, if 'finalSymbols' returns with any 'undefined' whatsoever, a message
about an impossible combination being set will be displayed instead of
the app window crashing.